Rename providers files.
