+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import { ChakraProvider } from '@chakra-ui/react'
 import type { AppProps } from 'next/app'
-import ApolloProvider from 'providers/Apollo'
-import UserAuthProvider from 'providers/User'
+import ApolloProvider from 'providers/ApolloProvider'
+import UserAuthProvider from 'providers/UserProvider'
 
 const MyApp = ({ Component, pageProps }: AppProps) => (
   <ApolloProvider>
